rename scan to build, factor model.ts out of markup.ts
